Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Are there anyone who will actually use this? No one is assigned to work on https://bugs.chromium.org/p/chromium/issues/detail?id=.

Made 740276 blocked on 755746 since it's basically a requirement for a smooth transition.

> IMHO, adding this dimension now is premature. It will be just another dimension no one uses.

Dimensions are now cheap, so it doesn't matter to have dimensions that are not yet used. I also prefer to add the dimension preemptively so people do not feel helpless to start working on it.

Patch-set: 1
Reviewer: Marc-Antoine Ruel <1000487@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>