Update patch set 2
Patch Set 2:

> > > 
> > > This change is about (a) while reducing load on workers that have a dedicated GPU.
> > 
> > So, part of the problem is that there's no way to say "workers with a dedicated GPU should only accept tasks specifically requesting that GPU"?
> > Maybe a better way to work around this then is to partition those workers into a different pool?
> 
> gpu:none? That's the whole point of this CL.

I think that's the confusion. 

You want to bin things into (dedicated gpus) and (integrated gpus or no gpus at all), right? And you want to ensure that tests that can run correctly against *either* bin only actually get the latter?

My concern is that having the integrated-gpu bots "lie" and claim to also have no gpu will cause more problems than it solves, because it'll be confusing that the properties don't actually match reality.

Thus, I was suggesting that rather than discriminate on gpu:None, discriminate on pool:gpu vs. pool:chrome. Because then we're not conflating what the tasks are *capable* of doing with which resources should be dedicated to specific types of tests.

Alternatively, another way to do this is to create a new gpu:integrated_or_none and have users use that instead, but I still feel like that's confusing things.

Of course, if the layout tests actually are sensitive to the difference between "integrated" and "none", this is all moot.

Patch-set: 2