Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Swarming bots should try and pick up the *most specific* matching tasks first.
> > 
> > IE If there is a task with "gpu: intel" and a task *without* the gpu dimension, a bot should pick up the "gpu: intel" task first.
> 
> That's not how the service works. This kind of transient behavior is specifically out of scope so the service can scale up seamlessly and one can instantly know how many bots can run a task.
> 
> This kind of behavior is purely optimistic and doesn't run in a saturated state: it breaks down under a saturated load because then the scheduler doesn't have a choice left, the task will run to the first available bot every single time. Swarming was designed to work well under saturated workload.

You case *fails* under saturated load, mine does not.

If we have bots which have the following;
 100 bots = gpu:A
 100 bots = gpu:B
 100 bots = gpu:C

Then we have tasks which fall into the following buckets;
 Task 1 - gpu:A
 Task 2 - gpu:B
 Task 3 - gpu:don't care

We thus have;
 Task 1 - can run on 100 bots
 Task 2 - can run on 100 bots
 Task 3 - can run on 300 bots

If we have 75 * Task 1s, 75 * Task 2s, 90 * Task 3s -- if we do random assignment;

 Tasks of type 3 ends up;
   -  30 tasks running on gpu:A
   -  30 tasks running on gpu:B
   - ~30 tasks running on gpu:C

 Tasks of type 1 end up;
   - 70 tasks running on gpu:A
   -  5 tasks which *can not schedule*

 Tasks of type 2 end up;
   - 70 tasks running on gpu:B
   -  5 tasks which *can not schedule*

  Thus we have;
   - 10 tasks which *can not schedule*
   - 70 *idle* bots

If you *always* assign tasks via the most specific dimension we end up with;
 
 Tasks of type 1 end up;
   - 75 tasks running on gpu:A

 Tasks of type 2 end up;
   - 75 tasks running on gpu:B

 Tasks of type 3 ends up;
   - 25 tasks running on gpu:A
   - 25 tasks running on gpu:B
   - 40 tasks running on gpu:C

  Thus we have *all* tasks scheduled.

If you are not convinced I can make this even clearer if we have a single gpu:A bot. This is also trivial to prove via a simple simulation.

The *only* way to deal with a pool of non-homogeneous dimensions and tasks with non-homogeneous requests is to assign tasks which are more specific *first*.

Tim 'mithro' Ansell

Patch-set: 2