Update patch set 2
Patch Set 2: Code-Review+1

> Patch Set 2:
> 
> > > > 
> > > > This change is about (a) while reducing load on workers that have a dedicated GPU.
> > > 
> > > So, part of the problem is that there's no way to say "workers with a dedicated GPU should only accept tasks specifically requesting that GPU"?
> > > Maybe a better way to work around this then is to partition those workers into a different pool?
> > 
> > gpu:none? That's the whole point of this CL.
> 
> I think that's the confusion. 
> 
> You want to bin things into (dedicated gpus) and (integrated gpus or no gpus at all), right? And you want to ensure that tests that can run correctly against *either* bin only actually get the latter?
> 
> My concern is that having the integrated-gpu bots "lie" and claim to also have no gpu will cause more problems than it solves, because it'll be confusing that the properties don't actually match reality.
> 
> Thus, I was suggesting that rather than discriminate on gpu:None, discriminate on pool:gpu vs. pool:chrome. Because then we're not conflating what the tasks are *capable* of doing with which resources should be dedicated to specific types of tests.
> 
> Alternatively, another way to do this is to create a new gpu:integrated_or_none and have users use that instead, but I still feel like that's confusing things.
> 
> Of course, if the layout tests actually are sensitive to the difference between "integrated" and "none", this is all moot.

It's been under consideration for some time to put the physical GPU bots into their own pool: http://crbug.com/700053 . This would cover machines like the Retina MacBook Pros (both NVIDIA and AMD GPU), the Linux and Windows NVIDIA GPU bots, and more esoteric configurations like the Windows AMD bots.

Up to this point I hesitated to make that move because it seemed a bit antisocial. It seemed to me that tests which don't care whether they have a GPU or not should be able to spread out and run on any available machine.

However, the recent problems with the layout tests sharding to any available Mac bot, not taking into consideration whether there was a GPU or not (or Retina or not), was causing pixel differences, and ultimately more problems than it was solving.

I think that this CL would improve the situation. It would allow treating the Mac Minis similarly to the Mac VMs, and allow those two classes of machine to be targeted identically. The layout tests will probably run similarly on these two classes of machine. If they don't, then we should either improve the tests so they can, or scale up one configuration to be able to handle the layout tests' load.

From a brief chat with maruel@, it sounds like putting the physical GPU hardware into its own pool will clarify things better, so I'm going to take that bug and work on it. Still, this CL would probably improve the situation, so I'm +1'ing it to show my support.

Patch-set: 2
Reviewer: Kenneth Russell <1105344@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1