Update patch set 1
Patch Set 1:

It's kind of iffy but that's the only sensible workaround I can think of. I haven't been able to reproduce deterministically this bug but I know many people experienced it. The cost of this code is nearly zero in practice so it's worth trying.

Patch-set: 1
Reviewer: Marc-Antoine Ruel <1000487@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false