Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Drive-by question: shall we really (incorrectly) assume UTF-8 everywhere?

We really do.

There's two parts here; the file names and the content. The file names must be in a common encoding. There's no way around that.

For content, I started enforcing UTF-8 on Rietveld many many years back when files encoded in GB-2312 started to sneak in. They blow everything up; code search, source analysis, etc. Any file in the repository that is not UTF-8 encoded must be considered a binary file, because this is what they are.

> You mentioned Swarming bots with LANG correctly set do not have problems, so is setting LANG on all bots a more appropriate fix? I'm not sure if some other OS-level stuff may break without correct LANG.

That's a technical detail for our Ubuntu that doesn't matter that much in practice and personally I don't want to spend time on. It's been like that for a long time.

Patch-set: 3
Reviewer: Marc-Antoine Ruel <1000487@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0