Update patch set 2
Patch Set 2:

It's live on staging but this exception only happens on prod, likely due to the amount of items to process.

I confirmed the code to at least does not break anything when running on staging as version 3292-b8d8f94-tainted-maruel, so it should be safe to deploy.

Patch-set: 2
Reviewer: Marc-Antoine Ruel <1000487@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0
Work-in-progress: false