Update patch set 1
Patch Set 1:

Yep, sorry, wanted to reply to m-a, but didn't finish new patch yet.

To summarize with m-a offline conversation:
  * make all "clear-out" semantics of merging be errors instead (envvars, env prefixes, cipd_package versions, caches)
  * keep inclusions otherwise (I believe that they are ~required to have non-insane config).

The rationale is to make the inclusion/merge logic as simple and restrictive as we can get away with for now. For templates which want to 'delete' a base layer, they'd be encouraged to instead split their includes to be more fine-grained instead.

Patch-set: 1
Reviewer: Robbie Iannucci <1001907@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>