Update patch set 9
Patch Set 9:

Good morning world :)

Can I get another round of review on this? I'd really like to get this all landed and deployed in the next handful of days, or it's going to be in limbo for a month+ (due to paternity leave).

Patch-set: 9
Reviewer: Robbie Iannucci <1001907@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>