Update patch set 7
Patch Set 7:

(2 comments)

So the potential downside here is that it makes it possible/easy to accidentally forget to call trim() somewhere, which could potentially lead to the cache getting bigger and bigger and somehow getting into a situation where it's not being trimmed, right?

When the cache trimming is moved to a separate process, will all of these calls to trim here be removed? How will we ensure that caches will always be regularly trimmed?

Patch-set: 7
Reviewer: Quinten Yearsley <1003190@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>