Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review+1
> > 
> > LGTM with questions ---
> > 1) how arbitrary is arbitrary? :)
> > 2) does it also make sense to put somewhere a note to check here if the "too many instance templates" error crops up again? Ideally there would be some sort of config validation to check that; perhaps it's encompassed by the changes you're planning.
> 
> I think it'd be too many steps and we should just remove this limit. GCE quota will prevent too many instance templates from being created. The limit of arbitrary here is whatever number of templates OOMs the server. I suspect a reviewer will see and question it if someone tries to add however many templates it'd take to do that (presumably reviewers can tell the difference between +1 template being necessary and +1000).

OK, sounds good, thanks for the clarification! I was wondering from the perspective of, someone without context might not know this file needed updating upon running up against the limit, so how might we communicate that information (I'd originally wondered if it was indeed a quota we were hitting, but couldn't find where that quota might live); if we remove the limit entirely this is moot.

Patch-set: 2