Update patch set 5
Patch Set 5:

Right now the biggest disadvantage of this approach is the poor schema created in BigQuery via bqschemaupdater. The result of RECORD (key, value:values) is suboptimal. That said, I think this can be fixed if we want to. If this can be fixed, then we need to think if it's a strict improvement.

Patch-set: 5
Reviewer: Gerrit User 1000487 <1000487@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0