Update patch set 3
Patch Set 3: Code-Review-1

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > As far as generalization goes, this would be nice to use.  I already have a hacky version of this for some of our SSH bots (Chromecast/ChromeOS) https://chrome-internal.googlesource.com/infradata/config/+/master/configs/chromium-swarm/scripts/skia_mobile.py#56
> > 
> > It would be nice to pipe the dimensions in the "right" way
> 
> +1 We have some custom bot logic that already does this as well. Might be nice to migrate to this method instead to reduce complexity from our side.


I don't like saying no, and this one deserve an (hopefully) clear explanation:


I still think that dropping configuration files on the host for the bot is a very strong anti-pattern.

This creates an implicit API; the bot starts, looks for a file, if found, it uses the content.

Next thing you want is for the bot to monitor for file change, and restart when it's modified.

Then you see you want to add new state X to the file (setting environment variable? filtering commands? doesn't matter) so you need to do a multi-partite update mechanism every time:
- commit new code that supports both the old and the new
- push that to prod
- push the new files, for each team, through their host maintenance mechanism (puppet, salt, etc)
- Once this inter-team coordination is complete, commit code to remove the old functionality
- push that to prod

This is very intensive for the Swarming team (me). Been there, done that.
Where having each team do it on their own in their own bot_config cost me nothing (yay diffuse costs) but the real reason is that if you need to change the format, I'm not in the blocking loop (higher velocity).

But then you tell me "M-A, the format won't change, you're freaking out". You are problably right, bot the bot itself will change. To implement https://crbug.com/638071 (Deploy bot_config as a CIPD package), we need to create an API to make it asynchronous. This dimension.json new functionality creates more questions: should the CIPD package ignore these file injected dimensions, ignore these? Should the package be the one reading theses?


Most importantly,

If you realized, the bot intentionally keeps extremely minimal state from the host. In practice, the only stateful content are the caches, and each caches can be safely deleted, they are not a *source of truth*.

This is important, because now this dimension.json becomes a source of truth; there are new failure modes:
- the file is missing
- the file is stale (was updated but the bot didn't notice)
- the file is corrupted
- salt/puppet failed to push, disk space issues, etc

You are now coordinating with another team: labs, and I'm in the loop for changes.

I'm a huge advocate of having the minimum of ways to do a single operation, *especially* for configuration changes (see recent code red, you don't want to have an L7+ review your single line change, right?).

All of the Chrome Swarming instances migrated to Starlark based configuration files, and anything that should be a forced dimension should be set there.


In conclusion, I don't think it hits the generalization bar because:

- It creates a versioning problem based on an unstructured format that have very high friction (low velocity to change to do multi-partite coordination) and is hard to future proof correctly.
- It creates a new way to do something that can already be done.
- It encourage users to push the source of truth to be stateful on the bots, versus being centralized, which is against the philosophy of the service.
- This will make migrating dimensions to be generated on CIPD harder than needed.

Patch-set: 3
Reviewer: Gerrit User 1000487 <1000487@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1