Update patch set 5
Abandoned

This was a mitigation, but the issue is currently mitigated. Because of that I'm not going to add the technical debt of the shardMap https://bugs.chromium.org/p/chromium/issues/detail?id=960437&q=owner%3Ame&colspec=ID%20Pri%20M%20Stars%20ReleaseBlock%20Component%20Status%20Owner%20Summary%20OS%20Modified

I'll put out a similar review but instead of a manual map of id ranges it will contain a new indexed datastore property on BotInfo that we can shard on.

Thanks for your feedback, it will be included in the next version.

Patch-set: 5
Status: abandoned
Tag: autogenerated:gerrit:abandon