Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review+1
> 
> One thing I wonder; since the task hasn't really started, I wish the task result would be CANCELED instead of KILLED. But in practice a but did get a bot assigned to it.
> 
> I'm not sure it's a problem worth fixing, so lgtm.

Thank you for your advice.

Indeed, it should be cancelled. And I think it's doable by returning a flag to the server. e.g. is_cancelled=True
But it might be a bit odd for user since the BotInfo.status was already updated to RUNNING at polling.

It makes sense if the BotInfo.status gets updated from PENDING to RUNNING when the Bot actually runs the task via task_runner.py. So I will consider the refactoring when it becomes a problem.

Does this make sense?

Patch-set: 4