Update patch set 5
Patch Set 5:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > even if this file is self-consistent, it is inconsistent with other files in the same package.
> > Does this CL has to combine both 1) new feature and 2) new code style?
> 
> I think it is fine if the package eventually be consistent after changes. More over, I don't think we need to discuss about format / code style while yapf has such responsibility.

Combining style changes and code changes in one CL makes reviewing harder. The reviewer has to read each hunk and see if there was a functional change. This increase the load on the reviewer, and looks like it will increase the load in future CLs. This is why it is better to separate CLs. In the CL with only style changes, the reviewer would be confident that all these changes were automatic and there is no need to carefully review each changed line. 

Consider sending a separate CL reformatting entire repo. It helped me in buildbucket: https://chromium-review.googlesource.com/c/infra/infra/+/1050499/

Patch-set: 5